Abstracted the toggle button animation to an isolated hook.
